remove unnecesary migrations
