+++ protocols/migration/__init__.py
@@ -6,10 +6,8 @@ from .migration_participant_100_to_participant_103 import MigrationParticipants1
 from .migration_participant_103_to_participant_110 import MigrationParticipants103To110
 from .migration_reports_300_to_participant_100 import MigrationReports3ToParticipant1
 from .migration_participant_101_to_reports_300 import MigrationParticipants101ToReports
-from .migration_participant_103_to_participant_100 import MigrationParticipants103To100
 from .migration_participant_100_to_reports_300 import MigrationParticipants100ToReports
 from .migration_participant_110_to_participant_100 import MigrateParticipant110To100
-from .migration_participant_110_to_participant_103 import MigrateParticipant110To103
 from .migration_reports_400_to_reports_300 import MigrateReports400To300
 
 from .migration_reports_300_to_reports_400 import MigrateReports3To4
