+++ protocols/tests/test_migration/test_migration_participant_1_1_0_to_participant_1_0_0.py
@@ -35,4 +35,11 @@ class TestMigrateParticipant110To100(TestCaseMigration):
         ]:
             self.assertEqual(getattr(pm_1_1_0, attribute), getattr(pm_1_0_0, attribute))
 
+    def test_migrate_cancer_participant(self):
+        p_1_1_0 = self.get_valid_object(
+            object_type=self.old_model.CancerParticipant, version=self.version_7_0, fill_nullables=False)
+        p_1_0_0 = MigrateParticipant110To100().migrate_cancer_participant(old_instance=p_1_1_0)
+        self.assertIsInstance(p_1_0_0, self.new_model.CancerParticipant)
+        self.assertTrue(p_1_0_0.validate(p_1_0_0.toJsonDict()))
+
